Moved ReferenceCounter to System.Buffers.Experimental
