+++ src/System.Buffers.Experimental/System/Runtime/ReferenceCounter.cs
