+++ src/System.Buffers.Experimental/System/Runtime/ReferenceCountingSettings.cs
