+++ src/System.Buffers.Primitives/System/Buffers/OwnedBuffer.cs
@@ -58,23 +58,4 @@ namespace System.Buffers
         bool _disposed;
         #endregion
     }
-
-    public abstract class ReferenceCountedBuffer<T> : OwnedBuffer<T>
-    {
-        int _referenceCount;
-
-        public override void AddReference()
-        {
-            Interlocked.Increment(ref _referenceCount);
-        }
-
-        public override void Release()
-        {
-            if (Interlocked.Decrement(ref _referenceCount) == 0) {
-                OnZeroReferences();
-            }
-        }
-
-        public override bool HasOutstandingReferences => _referenceCount > 0;
-    }
 }
