+++ tests/System.Buffers.Experimental.Tests/MemoryTests.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 using System.Buffers;
+using System.Collections.Generic;
+using System.Runtime;
 using Xunit;
 
 namespace System.Slices.Tests
